Implement PlaylistCard and playlist services
The update includes the addition of the new PlaylistCard component and playlist services. These changes allow the fetching and display of YouTube playlists, which can be found in a new file within the services folder, playlistService.ts. The PlaylistPage component was also updated to use these new features and display playlists.